enhancement(Calendar): make corrections and add a hooks folder and its custom webpack alias
